+++ node_server/node_server.py
@@ -1,139 +1,11 @@
-from hashlib import sha256
-import json
-import time
-from datetime import datetime
-
+from models.Blockchain import Blockchain
+from models.Blockchain.Block import Block
+from models.Blockchain.Block.Transaction import Transaction
+from helpers.decorator import convert_input_to
 from flask import Flask, request
+import jsons
+import time
 import requests
-
-
-class Block:
-    def __init__(self, index, transactions, datetime, previous_hash, nonce=0):
-        self.index = index
-        self.transactions = transactions
-        self.datetime = datetime
-        self.previous_hash = previous_hash
-        self.nonce = nonce
-
-    def compute_hash(self):
-        """
-        A function that return the hash of the block contents.
-        """
-        block_string = json.dumps(self.__dict__, sort_keys=True)
-        return sha256(block_string.encode()).hexdigest()
-
-
-class Blockchain:
-    # difficulty of our PoW algorithm
-    difficulty = 2
-
-    def __init__(self):
-        self.unconfirmed_transactions = []
-        self.chain = []
-
-    def create_genesis_block(self):
-        """
-        A function to generate genesis block and appends it to
-        the chain. The block has index 0, previous_hash as 0, and
-        a valid hash.
-        """
-        genesis_block = Block(0, [], 0, "0")
-        genesis_block.hash = genesis_block.compute_hash()
-        self.chain.append(genesis_block)
-
-    @property
-    def last_block(self):
-        return self.chain[-1]
-
-    def add_block(self, block, proof):
-        """
-        A function that adds the block to the chain after verification.
-        Verification includes:
-        * Checking if the proof is valid.
-        * The previous_hash referred in the block and the hash of latest block
-          in the chain match.
-        """
-        previous_hash = self.last_block.hash
-
-        if previous_hash != block.previous_hash:
-            return False
-
-        if not Blockchain.is_valid_proof(block, proof):
-            return False
-
-        block.hash = proof
-        self.chain.append(block)
-        return True
-
-    @staticmethod
-    def proof_of_work(block):
-        """
-        Function that tries different values of nonce to get a hash
-        that satisfies our difficulty criteria.
-        """
-        block.nonce = 0
-
-        computed_hash = block.compute_hash()
-        while not computed_hash.startswith('0' * Blockchain.difficulty):
-            block.nonce += 1
-            computed_hash = block.compute_hash()
-
-        return computed_hash
-
-    def add_new_transaction(self, transaction):
-        self.unconfirmed_transactions.append(transaction)
-
-    @classmethod
-    def is_valid_proof(cls, block, block_hash):
-        """
-        Check if block_hash is valid hash of block and satisfies
-        the difficulty criteria.
-        """
-        return (block_hash.startswith('0' * Blockchain.difficulty) and
-                block_hash == block.compute_hash())
-
-    @classmethod
-    def check_chain_validity(cls, chain):
-        result = True
-        previous_hash = "0"
-
-        for block in chain:
-            block_hash = block.hash
-            # remove the hash field to recompute the hash again
-            # using `compute_hash` method.
-            delattr(block, "hash")
-
-            if not cls.is_valid_proof(block, block_hash) or \
-                    previous_hash != block.previous_hash:
-                result = False
-                break
-
-            block.hash, previous_hash = block_hash, block_hash
-
-        return result
-
-    def mine(self):
-        """
-        This function serves as an interface to add the pending
-        transactions to the blockchain by adding them to the block
-        and figuring out Proof Of Work.
-        """
-        if not self.unconfirmed_transactions:
-            return False
-
-        last_block = self.last_block
-
-        new_block = Block(index=last_block.index + 1,
-                          transactions=self.unconfirmed_transactions,
-                          datetime=datetime.now().strftime("%Y/%m/%d %H:%M:%S"),
-                          previous_hash=last_block.hash)
-
-        proof = self.proof_of_work(new_block)
-        self.add_block(new_block, proof)
-
-        self.unconfirmed_transactions = []
-
-        return True
 
 
 app = Flask(__name__)
@@ -150,25 +22,7 @@ peers = set()
 # our application to add new data (posts) to the blockchain
 @app.route('/new_transaction', methods=['POST'])
 def new_transaction():
-    tx_data = request.get_json()
-    contents = []
-
-    if(tx_data.get('type') == 'inscription'):
-        contents = ['name', 'email']
-
-    if(tx_data.get('type') == 'transaction'):
-        contents = ['text']
-
-    if not tx_data.get('user_name'):
-        return "invalid transaction data", 404
-
-    if (len(contents) > 0):
-        for cont in contents:
-            if not tx_data.get('content').get(cont):
-                return "invalid transaction data", 404
-
-    blockchain.add_new_transaction(tx_data)
-
+    blockchain.add_new_transaction(Transaction.from_json(request.get_json()))
     return "Success", 201
 
 
@@ -179,8 +33,8 @@ def new_transaction():
 def get_chain():
     chain_data = []
     for block in blockchain.chain:
-        chain_data.append(block.__dict__)
-    return json.dumps({"length": len(chain_data),
+        chain_data.append(block)
+    return jsons.dumps({"length": len(chain_data),
                        "chain": chain_data,
                        "peers": list(peers)})
 
@@ -234,7 +88,7 @@ def register_with_existing_node():
 
     # Make a request to register with remote node and obtain information
     response = requests.post(node_address + "/register_node",
-                             data=json.dumps(data), headers=headers)
+                             data=jsons.dumps(data), headers=headers)
 
     if response.status_code == 200:
         global blockchain
@@ -291,7 +145,7 @@ def verify_and_add_block():
 # endpoint to query unconfirmed transactions
 @app.route('/pending_tx')
 def get_pending_tx():
-    return json.dumps(blockchain.unconfirmed_transactions)
+    return jsons.dumps(blockchain.unconfirmed_transactions)
 
 
 def consensus():
@@ -329,7 +183,7 @@ def announce_new_block(block):
         url = "{}add_block".format(peer)
         headers = {'Content-Type': "application/json"}
         requests.post(url,
-                      data=json.dumps(block.__dict__, sort_keys=True),
+                      data=jsons.dumps(block.__dict__, sort_keys=True),
                       headers=headers)
