+++ node_server/requirements.txt
@@ -1,3 +1,5 @@
 Flask~=1.1
 requests~=2.22
-
+jsons==1.2.0
+SQLAlchemy==1.3.19
+flask-jsontools==0.1.7
